% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemotherapyRecordsExtraction.R
\name{chemotherapyRecordsExtraction}
\alias{chemotherapyRecordsExtraction}
\title{Extract chemotherapy records for target regimen.}
\usage{
chemotherapyRecordsExtraction(targetParameter, connection, cohortTable,
  cdmDatabaseSchema, cohortDatabaseSchema, targetCohortId, maxCores)
}
\arguments{
\item{maxCores}{Number of cores in clusterApply}
}
\value{
chemotherapy records for specific regimen
}
\description{
Extract chemotherapy records for target regimen.
}
\keyword{chemotherapy}
\keyword{records}
