% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemotherapyToEpisode.R
\name{chemotherapyToEpisode}
\alias{chemotherapyToEpisode}
\title{Transform regimen records to episode table.}
\usage{
chemotherapyToEpisode(chemotherapyRecords)
}
\value{
list of episode table and episode event table
}
\description{
Transform regimen records to episode table.
}
