% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGeneration.R
\name{createCohort}
\alias{createCohort}
\title{Create cohort of interest by using condition concept Ids}
\usage{
createCohort(createCohortTable = F, connectionDetails,
  oracleTempSchema = NULL, cdmDatabaseSchema, cohortDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema, cohortTable,
  conceptIdSet = c(), includeConceptIdSetDescendant = F,
  targetCohortId)
}
\arguments{
\item{createCohortTable}{}

\item{connectionDetails}{}

\item{oracleTempSchema}{}

\item{cdmDatabaseSchema}{}

\item{cohortDatabaseSchema}{}

\item{vocabularyDatabaseSchema}{}

\item{cohortTable}{}

\item{conceptIdSet}{}

\item{includeConceptIdSetDescendant}{}

\item{targetCohortId}{}
}
\description{
Create cohort of interest by using condition concept Ids
}
