% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTargetCohort.R
\name{createCohortTable}
\alias{createCohortTable}
\title{Create cohort table, Insert target cohort}
\usage{
createCohortTable(connection, oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema, cohortTable)
}
\description{
Create cohort table, Insert target cohort
}
