% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEpisodeTable.R
\name{createEpisodeTable}
\alias{createEpisodeTable}
\title{Create Episode / Episode_event table}
\usage{
createEpisodeTable(connection, oracleTempSchema, oncologyDatabaseSchema,
  episodeTable, episodeEventTable)
}
\description{
Create Episode / Episode_event table
}
