% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractionAlgorithm.R
\name{DrugExposureInCohort}
\alias{DrugExposureInCohort}
\title{Algorithm for chemotherapy record extraction}
\usage{
DrugExposureInCohort(
  targetConceptIds,
  connectionDetails,
  cohortTable,
  includeDescendant = TRUE,
  outofCohortPeriod = TRUE,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  targetCohortId
)
}
\description{
Algorithm for chemotherapy record extraction
}
