% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Excute}
\usage{
execute(connectionDetails, connection, cohortTable, includeDescendant,
  outofCohortPeriod, cohortDatabaseSchema, primaryDrugList,
  secondaryDrugList, eliminatoryDrugList, targetCohortId,
  createCsv = FALSE, regimenName = "Unknown")
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}
}
\value{
histogram and count for subject distribution in repeated cycle,The number of distinct person_id in result, episode table in CSV file
}
\description{
Excute the Study
}
\examples{
execute(connectionDetails,
connection,
cohortTable = cohortTable,
includeDescendant = includeDescendant,
outofCohortPeriod = outofCohortPeriod,
cohortDatabaseSchema = cohortDatabaseSchema,
primaryDrugList = primaryDrugList,
secondaryDrugList = secondaryDrugList,
eliminatoryDrugList= eliminatoryDrugList,
targetCohortId = targetCohortId,
createCsv = createCsv
regimenName = 'FOLFOX')
}
\keyword{execute}
