% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{extractTargetRegimen}
\alias{extractTargetRegimen}
\title{Populating chemotherapy regimens}
\usage{
extractTargetRegimen(parameters, connectionDetails, cohortTable,
  cdmDatabaseSchema, cohortDatabaseSchema, targetCohortId, maxCores)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
                         \code{\link[DatabaseConnector]{createConnectionDetails}} function in the
                         DatabaseConnector package.
Excute

Excute the Study}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{targetRegimenConceptIds}{The concept id list of the regimen}
}
\value{
Generated episode table
}
\description{
Populating chemotherapy regimens
}
\examples{
generateEpisodeTable(targetRegimenConceptIds,
                    connectionDetails,
                    cohortTable,
                    cdmDatabaseSchema,
                    cohortDatabaseSchema,
                    targetCohortId,
                    maxCores)
}
\keyword{generateEpisodeTable}
