% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GapDateExamination.R
\name{gapDateExamination}
\alias{gapDateExamination}
\title{gapDateExamination}
\usage{
gapDateExamination(subjectId)
}
\arguments{
\item{subjectId}{: targeting subject_id in target cohort}
}
\value{
cycle result list of single subject in cohort
}
\description{
check cycle interval in the range of setting
}
\examples{
gapDateExamination(subjectId = 11111111)
}
\keyword{date}
\keyword{gap}
