% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateEpisode.R
\name{generateEpisode}
\alias{generateEpisode}
\title{Generate chemotherapy records in episode table and insert to database.}
\usage{
generateEpisode(parameters, connection, cohortTable, cdmDatabaseSchema,
  cohortDatabaseSchema, targetCohortId, maxCores)
}
\arguments{
\item{maxCores}{Number of cores using in clusterApply}
}
\value{
Episode / Episode_event table
}
\description{
Generate chemotherapy records in episode table and insert to database.
}
\keyword{Episode}
\keyword{Extraction,}
