% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{generateEpisodeTable}
\alias{generateEpisodeTable}
\title{GenerateEpisodeTable
Generate regimen records for target regimen in episode table form and insert to database.}
\usage{
generateEpisodeTable(targetRegimenConceptIds, connectionDetails,
  cohortTable, cdmDatabaseSchema, cohortDatabaseSchema, targetCohortId,
  maxCores)
}
\arguments{
\item{targetRegimenConceptIds}{regimen concept ids}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{cdmDatabaseSchema}{}

\item{maxCores}{Number of cores using in clusterApply}
}
\value{
records of the target single regimen
}
\description{
GenerateEpisodeTable
Generate regimen records for target regimen in episode table form and insert to database.
}
\keyword{records}
\keyword{regimen,}
\keyword{target}
