% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSetting.R
\name{listForm}
\alias{listForm}
\title{Parameter setting for algorithm.}
\usage{
listForm(regimenInstance, targetRegimenConceptId)
}
\value{
parameters
}
\description{
Parameter setting for algorithm.
}
\examples{
parameterSetting(targetRegimenConceptIds=targetRegimenConceptIds)
}
\keyword{parameter}
