% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimenSetting.R
\name{parameterSetting}
\alias{parameterSetting}
\title{regimenSetting}
\usage{
parameterSetting(jsonName = "regimenDrugSettingJsonForm.json",
  targetRegimenConceptIds)
}
\value{
drugList
}
\description{
regimenSetting
}
\examples{
regimenSetting(connectionDetails = connectionDetails,
connection = connection,
vocaDatabaseSchema = connectionDetails$schema,
regimenConceptId = regimenConceptId)
}
\keyword{regimenSetting}
