% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSetting.R
\name{readJson}
\alias{readJson}
\title{ParameterSetting
Parameter setting for algorithm.}
\usage{
readJson(jsonName = "regimenDrugSettingJsonForm.json")
}
\value{
parameters
}
\description{
ParameterSetting
Parameter setting for algorithm.
}
\examples{
parameterSetting(targetRegimenConceptIds=targetRegimenConceptIds)
}
\keyword{Setting}
\keyword{regimen}
