% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEpisodeTable.R
\name{recordsInEpisodeTableForm}
\alias{recordsInEpisodeTableForm}
\title{regimen records to episode table}
\usage{
recordsInEpisodeTableForm(regimenRecords, maxEpisodeId)
}
\arguments{
\item{regimenRecords}{regimen records}
}
\value{
list of episode table and episode event table
}
\description{
regimen records to episode table
}
\examples{
recordsInEpisodeTableForm(regimenRecords)
}
\keyword{episode}
