% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepetitionTrendHeatmap.R
\name{regimenHeatmap}
\alias{regimenHeatmap}
\title{Heatmap
Trends in regimen heatmap}
\usage{
regimenHeatmap(episodeTableFromDatabase,
  visualizationTargetRegimenId = NULL, heatmapInRatio = TRUE,
  maximumCycleNumber = NULL)
}
\value{
repitition trend heatmap
}
\description{
Heatmap
Trends in regimen heatmap
}
\keyword{heatmap}
