% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimenSetting.R
\name{regimenSetting}
\alias{regimenSetting}
\title{regimenSetting}
\usage{
regimenSetting(connectionDetails, connection,
  vocaDatabaseSchema = connectionDetails$schema, regimenConceptId,
  indexDrugConceptId)
}
\value{
drugList
}
\description{
regimenSetting
}
\examples{
regimenSetting()
}
\keyword{regimenSetting}
