% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleEditor.R
\name{ruleEditor}
\alias{ruleEditor}
\title{RuleEditor
Edit JSON for algorithm.}
\usage{
ruleEditor(new = FALSE, targetRegimenIds = NULL)
}
\value{
new JSON file
}
\description{
RuleEditor
Edit JSON for algorithm.
}
\keyword{JSON,edit}
