% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SankeyDiagram.R
\name{surgeryForSankey}
\alias{surgeryForSankey}
\title{sankey for regimen and surgery
Visualization tool for sankey for regimen and surgery}
\usage{
surgeryForSankey(connectionDetails, cdmDatabaseSchema,
  oncologyDatabaseSchema, episodeTable, surgeryConceptId)
}
\value{
sankey for regimen and surgery
}
\description{
sankey for regimen and surgery
Visualization tool for sankey for regimen and surgery
}
\keyword{sankey}
