% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeutrophilAnalysis.R
\name{treatmentLineFromEpisode}
\alias{treatmentLineFromEpisode}
\title{NeutrophilGraph
Visualization tool for episode table.}
\usage{
treatmentLineFromEpisode(connectionDetails, vocaDatabaseSchema,
  oncologyDatabaseSchema, episodeTable)
}
\value{
Graph or data for neutrophil analysis
}
\description{
NeutrophilGraph
Visualization tool for episode table.
}
\keyword{neutrophil,visualization}
