% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionTable.R
\name{distributionTable}
\alias{distributionTable}
\title{generate regimen records for target regimen}
\usage{
distributionTable(episodeTable, episodeSourceConceptId)
}
\arguments{
\item{targetRegimenConceptIds}{regimen concept ids}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{cdmDatabaseSchema}{}

\item{maxCores}{Number of cores using in clusterApply}
}
\value{
records of the target single regimen
}
\description{
generate regimen records for target regimen
}
\keyword{records}
\keyword{regimen,}
\keyword{target}
