% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{distributionTable}
\alias{distributionTable}
\title{Visualization
Visualization tool for episode table.}
\usage{
distributionTable(episodeTable, targetEpisodeConceptId)
}
\value{
Graph or table for episode table
}
\description{
Visualization
Visualization tool for episode table.
}
\keyword{graph,visualization}
