% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugExposureInCohort.R
\name{DrugExposureInCohort}
\alias{DrugExposureInCohort}
\title{DrugExposureInCohort}
\usage{
DrugExposureInCohort(connectionDetails, connection, cohortTable,
  includeDescendant = TRUE, outofCohortPeriod = TRUE,
  cohortDatabaseSchema, targetConceptIds, targetCohortId)
}
\value{
list form drug exposure data list
}
\description{
This function allows you to extract drug exposure records of target cohort
}
\examples{
DrugListinCohort(connectionDetails,connection,cohortTable,includeDescendant,outofCohortPeriod,cohortDatabaseSchema,drugList,targetCohortId)
}
\keyword{drug}
\keyword{list}
