% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugExposureInCohort.R
\name{DrugExposureInCohort}
\alias{DrugExposureInCohort}
\title{Drug exposure records}
\usage{
DrugExposureInCohort(targetConceptIds, connection, cohortTable,
  includeDescendant = TRUE, outofCohortPeriod = TRUE,
  cdmDatabaseSchema, cohortDatabaseSchema, targetCohortId)
}
\value{
Drug exposure
}
\description{
Drug exposure records
}
\keyword{Drug}
