% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugRecordExamination.R
\name{drugRecordExamination}
\alias{drugRecordExamination}
\title{drugRecordExamination}
\usage{
drugRecordExamination(targetSubjectId, primaryConceptRecords,
  secondaryConceptRecords, excludingConceptRecords, drugInspectionDate,
  secondaryConceptIdList, excludingConceptIdList)
}
\arguments{
\item{subjectId}{: targeting subjectId in target cohort}
}
\value{
drug condition passed index dates list
}
\description{
check each index dates of single subject whether drug conditions are satisfied or not.
}
\examples{
drugRecordExamination(subjectId = 11111111)
}
\keyword{primaryDrug,}
\keyword{secondaryDrug,eliminatoryDrug}
