% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugRecordExamination.R
\name{drugRecordExamination}
\alias{drugRecordExamination}
\title{DrugRecordExamination
check each index dates of single subject whether drug conditions are satisfied or not.}
\usage{
drugRecordExamination(targetSubjectId, primaryConceptRecords,
  secondaryConceptRecords, excludingConceptRecords, drugInspectionDate,
  secondaryConceptIdList, excludingConceptIdList)
}
\value{
drug condition passed dates list
}
\description{
DrugRecordExamination
check each index dates of single subject whether drug conditions are satisfied or not.
}
\examples{
drugRecordExamination(subjectId = 11111111)
}
\keyword{primaryDrug,}
\keyword{secondaryDrug,eliminatoryDrug}
