% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTargetRegimen.R
\name{extractTargetRegimen}
\alias{extractTargetRegimen}
\title{ExtractTargetRegimen
Generate regimen records for target regimen.}
\usage{
extractTargetRegimen(parameters, connectionDetails, cohortTable,
  cdmDatabaseSchema, cohortDatabaseSchema, targetCohortId, maxCores)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{maxCores}{Number of cores using in clusterApply}
}
\value{
The records of the target single regimen
}
\description{
ExtractTargetRegimen
Generate regimen records for target regimen.
}
\keyword{records}
\keyword{regimen,}
\keyword{target}
