% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GapDateExamination.R
\name{gapDateExamination}
\alias{gapDateExamination}
\title{GapDateExamination
check cycle interval in the range of setting}
\usage{
gapDateExamination(targetSubjectId, primaryConceptRecords,
  secondaryConceptRecords, excludingConceptRecords, drugInspectionDate,
  secondaryConceptIdList, excludingConceptIdList, gapDateBetweenCycle,
  gapDateBefore, gapDateAfter, regimenConceptId)
}
\value{
Cycle list of single subject in cohort
}
\description{
GapDateExamination
check cycle interval in the range of setting
}
\keyword{date}
\keyword{gap}
