% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimenSetting.R
\name{readJson}
\alias{readJson}
\title{regimenSetting}
\usage{
readJson(jsonName = "regimenDrugSettingJsonForm.json")
}
\value{
drugList
}
\description{
regimenSetting
}
\examples{
parameterSetting(targetRegimenConceptIds=targetRegimenConceptIds)
}
\keyword{regimenSetting}
