% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSetting.R, R/ParameterTree.R
\name{readJson}
\alias{readJson}
\title{ParameterSetting
Parameter setting for algorithm.}
\usage{
readJson(jsonName = "regimenDrugSettingJsonForm.json")

readJson(jsonName = "regimenDrugSettingJsonForm.json")
}
\value{
parameters

Graph or table for episode table
}
\description{
ParameterSetting
Parameter setting for algorithm.

Visualization
Visualization tool for episode table.
}
\examples{
parameterSetting(targetRegimenConceptIds=targetRegimenConceptIds)
}
\keyword{Setting}
\keyword{graph,visualization}
\keyword{regimen}
