% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordsInEpisodeTableForm.R
\name{recordsInEpisodeTableForm}
\alias{recordsInEpisodeTableForm}
\title{RecordsInEpisodeTableForm
Transform regimen records into episode table form.}
\usage{
recordsInEpisodeTableForm(regimenRecords)
}
\arguments{
\item{regimenRecords}{regimen records}
}
\value{
list of episode table and episode event table
}
\description{
RecordsInEpisodeTableForm
Transform regimen records into episode table form.
}
\examples{
recordsInEpisodeTableForm(regimenRecords)
}
\keyword{episode}
